+++ RELEASES.md
@@ -12,6 +12,8 @@ The TODO section is part of the [README.md](https://github.com/bestia-dev/counte
 
 - auto
 
+- update automation, clippy
+
 ## Version 2024.1129.1419 (2024-11-29)
 
 - docs